+++ Program.cs
@@ -1,9 +1,10 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
+using LabManager.repositories;
 
 new DataBaseSetup();
 
-
+var computerRepository = new ComputerRepository();
 
 //routing
 var modelName = args[0];
@@ -15,22 +16,11 @@ if(modelName == "Computer")
     if(modelAction == "List")
     {
         Console.WriteLine("Computer List");
-        var connection = new SqliteConnection("Data Source=database.db");
-       connection.Open();
-
-       var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers;"; 
-        
-        var reader = command.ExecuteReader();
-
-        while(reader.Read())
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine("{0},{1},{2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2)
-            );
+            Console.WriteLine("{0},{1},{2}", computer.Id, computer.Ram, computer.Processor);
         }
-
-        reader.Close();
-        connection.Close();
+       
     }
 
     if(modelAction == "New")
